refactor(multi-hop): extract MultiHopChainService — split plan vs execute (max-params 2/3)
MultiHopService injected 4 deps (search, planner, synthesizer, metrics).
Split on the plan-vs-execute seam: MultiHopChainService (search,
synthesizer, metrics) owns the whole execution — single-hop short-circuit,
the sequential hop loop with combination, runHop, the planner-outage
fallback, grounded synthesis, and all multi-hop metrics. MultiHopService
(planner, chain) produces the plan and delegates. Shared types + the
collectFactIds helper moved to multi-hop.types.ts to avoid a cycle.

The split dropped execute() below the complexity gate, so its
eslint-disable complexity directive is gone. Public run() contract
unchanged.

Part of the max-params=3 god-class split program.
