+++ src/multi-hop/multi-hop.module.ts
@@ -3,12 +3,13 @@ import { SearchModule } from '../search/search.module';
 import { SynthesizeModule } from '../synthesize/synthesize.module';
 import { MultiHopController } from './multi-hop.controller';
 import { MultiHopService } from './multi-hop.service';
+import { MultiHopChainService } from './multi-hop-chain.service';
 import { MultiHopPlannerService } from './multi-hop-planner.service';
 
 @Module({
   imports: [SearchModule, SynthesizeModule],
   controllers: [MultiHopController],
-  providers: [MultiHopService, MultiHopPlannerService],
+  providers: [MultiHopService, MultiHopChainService, MultiHopPlannerService],
   exports: [MultiHopService],
 })
 export class MultiHopModule {}
